Continuing Image Storage Work

+++ db/migrate/20231116182224_drop_images.rb
@@ -1,5 +1,5 @@
 class DropImages < ActiveRecord::Migration[7.0]
   def change
-    drop_table :images
+    drop_table :images if table_exists? :images
   end
 end
